# Stop generating whole module optimization flag