--- a/Generator/Sources/SwiftWinRT/writeProjectionFiles.swift
+++ b/Generator/Sources/SwiftWinRT/writeProjectionFiles.swift
@@ -25,21 +25,6 @@ internal func writeProjectionFiles(
             path: "\(directoryPath)\\CMakeLists.txt",
             directoryCreation: .ancestors))
 
-        // Generated modules can have a large number of files and are not manually edited
-        // so do not benefit from incremental compilation.
-        // By default, when building in debug mode, the Swift driver launches multiple swift-frontend processes,
-        // each of which reads every file in the module, a mode which enables incremental compilation
-        // but can result in quadratic compilation time.
-        // Whole module optimization uses a single swift-frontend process and reads every file only once,
-        // but does not allow incremental compilation.
-        // See: https://github.com/swiftlang/swift/blob/main/docs/CompilerPerformance.md#compilation-modes
-        let lineGroup = writer.output.createLineGroup()
-        writer.writeSingleLineCommand(lineGroup: lineGroup, "if", "POLICY", "CMP0157") // Swift_COMPILATION_MODE support
-        writer.writeSingleLineCommand(lineGroup: lineGroup, "set_directory_properties", "PROPERTIES", "Swift_COMPILATION_MODE", "wholemodule")
-        writer.writeSingleLineCommand(lineGroup: lineGroup, "elseif", .quoted("${CMAKE_BUILD_TYPE}"), "STREQUAL", "Debug")
-        writer.writeSingleLineCommand(lineGroup: lineGroup, "add_compile_options", "-whole-module-optimization")
-        writer.writeSingleLineCommand(lineGroup: lineGroup, "endif")
-
         // Workaround for https://github.com/swiftlang/swift-driver/issues/1477
         // The threshold value has to be real high because the driver multiplies the number of input files by ~50.
         // See https://github.com/swiftlang/swift-driver/blob/6af4c7dbc0559694578e5221d49970f94603b9e5/Sources/SwiftDriver/Jobs/FrontendJobHelpers.swift#L714
@@ -158,24 +143,6 @@ fileprivate func writeSwiftModuleFiles(
                 "set_target_properties", .autoquote(targetName),
                 "PROPERTIES", "Swift_MODULE_NAME", .autoquote(module.name))
         }
-
-        // The generated modules can have a large number of files and are not manually edited
-        // so do not benefit from incremental compilation.
-        // By default, when building in debug mode, the Swift driver launches multiple swift-frontend processes,
-        // each of which reads every file in the module, a mode which enables incremental compilation
-        // but can result in quadratic compilation time.
-        // Whole module optimization uses a single swift-frontend process and reads every file only once,
-        // but does not allow incremental compilation.
-        // See: https://github.com/swiftlang/swift/blob/main/docs/CompilerPerformance.md#compilation-modes
-        writer.writeSingleLineCommand("if", "POLICY", "CMP0157") // Swift_COMPILATION_MODE support
-        writer.writeSingleLineCommand(
-            "set_target_properties", .autoquote(targetName),
-            "PROPERTIES", "Swift_COMPILATION_MODE", "wholemodule")
-        writer.writeSingleLineCommand("elseif", .quoted("${CMAKE_BUILD_TYPE}"), "STREQUAL", "Debug")
-        writer.writeSingleLineCommand(
-            "target_compile_options", .autoquote(targetName),
-            "PRIVATE", .unquoted("-whole-module-optimization"))
-        writer.writeSingleLineCommand("endif")
 
         // Workaround for https://github.com/swiftlang/swift-driver/issues/1477
         // The threshold value has to be real high because the driver multiplies the number of input files by ~50.