It did not clearly make a difference:

```
With wmo
swift-winrt: 306.0073601 seconds
swift-winui-sample: 1299.6790674 seconds

With partial wmo
swift-winrt: 284.5574907 seconds
swift-winui-sample: 1188.8929425 seconds

Without wmo
swift-winrt: 336.6592267 seconds
swift-winui-sample: 1196.5236409 seconds
```